requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pdfplumber==0.10.3
pytesseract==0.3.10
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
email-validator==2.1.0
openai==1.51.0
openai==1.51.0
google-cloud-vision==3.4.5
